Add permissions persistence
