+++ lib/mumukit/auth/store.rb
@@ -1,28 +1,29 @@
 module Mumukit::Auth
   class Store
-    def initialize(db_name)
-      @db = Daybreak::DB.new "#{db_name}.db", default: '{}'
-    end
-
-    def close
-      @db.close
+    def initialize
+      @db = Mumukit::Auth.config.persistence_strategy
     end
 
     def set!(key, value)
-      @db.update! key.to_sym => value.to_json
+      @db.set! key.to_sym, value
     end
 
     def get(key)
-      Mumukit::Auth::Permissions.load @db[key]
+      @db.get key
+    end
+
+    def clean_env!
+      @db.clean_env!
     end
 
     class << self
+
       def from_env
-        new Mumukit::Auth.config.daybreak_name
+        Mumukit::Auth.config.persistence_strategy.class.from_env
       end
 
       def clean_env!
-        FileUtils.rm ["#{Mumukit::Auth.config.daybreak_name}.db"], force: true
+        from_env.clean_env!
       end
 
       def with(&block)
